+++ src/main/java/com/flab/gettoticket/entity/Users.java
@@ -1,4 +1,4 @@
-package com.flab.gettoticket.model;
+package com.flab.gettoticket.entity;
 
 import lombok.Getter;
 import lombok.Setter;
