--- src/main/java/com/flab/gettoticket/model/Zone.java
@@ -1,15 +1,0 @@
-package com.flab.gettoticket.model;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class Zone {
-    private long id;
-    private String grade;
-    private String name;
-    private int price;
-}
